@@ -1,22 +1,20 @@
 {
-  "name": "with-typescript",
+  "name": "with-typescript-plugin",
   "version": "1.0.0",
   "scripts": {
-    "dev": "concurrently \"tsc --pretty --watch\" \"next\"",
-    "prebuild": "tsc",
+    "dev": "next",
     "build": "next build",
     "start": "next start"
   },
   "dependencies": {
-    "next": "latest",
-    "react": "^16.1.0",
-    "react-dom": "^16.1.0"
+    "next": "^5.0.0",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0"
   },
   "devDependencies": {
-    "@types/next": "^2.4.5",
-    "@types/react": "^16.0.22",
-    "concurrently": "^3.5.0",
-    "tslint": "^5.8.0",
-    "typescript": "^2.6.1"
+    "@types/next": "^2.4.7",
+    "@types/react": "^16.0.36",
+    "@zeit/next-typescript": "0.0.8",
+    "typescript": "^2.7.1"
   }
 }
