@@ -11,7 +11,7 @@ export default function withRouter (ComposedComponent) {
       router: PropTypes.object
     }
 
-    static displayName = `withRoute(${displayName})`
+    static displayName = `withRouter(${displayName})`
 
     render () {
       const props = {
