@@ -106,10 +106,28 @@ export default function ({ app }, suiteName, render, fetch) {
       expect($('.as-path-content').text()).toBe('/nav/as-path?aa=10')
     })
 
-    test('error 404', async () => {
-      const $ = await get$('/non-existent')
-      expect($('h1').text()).toBe('404')
-      expect($('h2').text()).toBe('This page could not be found.')
+    describe('404', () => {
+      it('should 404 on not existent page', async () => {
+        const $ = await get$('/non-existent')
+        expect($('h1').text()).toBe('404')
+        expect($('h2').text()).toBe('This page could not be found.')
+      })
+
+      it('should 404 for <page>/', async () => {
+        const $ = await get$('/nav/about/')
+        expect($('h1').text()).toBe('404')
+        expect($('h2').text()).toBe('This page could not be found.')
+      })
+
+      it('should should not contain a page script in a 404 page', async () => {
+        const $ = await get$('/non-existent')
+        $('script[src]').each((index, element) => {
+          const src = $(element).attr('src')
+          if (src.includes('/non-existent')) {
+            throw new Error('Page includes page script')
+          }
+        })
+      })
     })
 
     describe('with the HOC based router', () => {
